float arguments accepted
sqr.py now takes float in addition to integer arguments

+++ sqr.py
@@ -1,11 +1,12 @@
 #! /usr/local/bin/python3.9
 # change above shebang line to reflect absolute path to preferred python interpreter
+# takes single integer/float value and displays its square to stdout
 
 import argparse
 
 parser = argparse.ArgumentParser()
 
-parser.add_argument("square", type=int, help="display square of given number")
+parser.add_argument("square", type=float, help="display square of given number")
 parser.add_argument("-v", "--verbose", type=int, help="increase output verbosity", choices=[0, 1, 2])
 
 args = parser.parse_args()
